requirements.txt
mysql-connector-python==9.1.0
seaborn==0.13.2
scikit-learn==1.5.1
wordcloud==1.9.3
PyPDF2==3.0.1
python-docx==1.0.1
pandas==2.1.1
matplotlib==3.8.0
